Began to modularize my scripts
