+++ scripts_nicholas/CountUniqueElementsColumn.py
@@ -6,14 +6,17 @@
 
 #------------------------------------------------------------------------------------------------
 import sys # For command-line arguments
+sys.path.append('/Users/nicholaslenz/Desktop/Summer2018 (Repo)/scripts_nicholas')
+import MiscFunctions as mf
 #------------------------------------------------------------------------------------------------
 
-fileRead = open(sys.argv[1], "r") # Opens given file for reading.
+fileRead = open(sys.argv[1], "r")
+column_number = mf.get_int_answer('What column should be accumulated? ')
 
 variants = []
 for variant in fileRead: # Adds every string in a given column to the list variants.
 	parsedVariant = variant.split('\t')
-	variants.append(parsedVariant[int(sys.argv[2])])
+	variants.append(parsedVariant[column_number])
 
 fileRead.close()
 print (len(set(variants))) # Construction of the set removes the duplicates in variants.
